renamed original server file to be replaced
